server/store_hunts/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
bcrypt==4.1.2
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.4
django-cors==0.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-env==0.2.1
django-rename-app==0.1.7
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
executing==2.0.1
filelock==3.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
ipython==8.24.0
isort==5.13.2
itemadapter==0.9.0
itemloaders==1.2.0
jedi==0.19.1
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.2.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
parsel==1.9.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
Protego==0.3.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
referencing==0.35.0
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tldextract==5.1.2
traitlets==5.14.3
Twisted==24.3.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
w3lib==2.1.2
wcwidth==0.2.13
zope.interface==6.3
